LPS-186666 Extract service from PortalCORSServletFilter to PortalCORSManagedServiceFactory, create PortalCORSRegistryUtil and apply to usages
